Update AI functionality
* changed how debug renders tiles
* changed how AI moves: now moves by center of tile instead of bottom left corner

+++ AsteroidsNetbeansModules/CommonEnemy/src/main/java/dk/sdu/mmmi/cbse/common/enemy/Enemy.java
@@ -1,26 +1,22 @@
 package dk.sdu.mmmi.cbse.common.enemy;
 
-import com.badlogic.gdx.math.Circle;
 import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.Graph;
-import dk.sdu.mmmi.cbse.common.data.Node;
-import dk.sdu.mmmi.cbse.common.data.Path;
 
 public class Enemy extends Entity {
     
-    private int aggroRange;
+    private float aggroRange;
     
-    public Enemy(int aggroRange) {
+    public Enemy(float aggroRange) {
         this.aggroRange = aggroRange;
         assignTexture("/img/enemy.png");
         setType("enemy");
     }
     
-    public void setAggroRange(int aggroRange) {
+    public void setAggroRange(float aggroRange) {
         this.aggroRange = aggroRange;
     }
 
-    public int getAggroRange() {
+    public float getAggroRange() {
         return aggroRange;
     }
